In this commit is adding the endpoint UPDATE IN HOSPITAL #1018

+++ src/main/java/es/upm/miw/apaw_practice/adapters/mongodb/Hospital/entities/DoctorEntity.java
@@ -9,7 +9,7 @@ public class DoctorEntity {
     @Id
     private String dni;
     private String fullname;
-    private Double salary;
+    private BigDecimal  salary;
 
     // Constructor
     public DoctorEntity(String dni, String fullname, Double salary) {
